Task 2 - implemented second power of number.

+++ src/main/java/cz/spsmb/hercik/gitTest/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Hello World!");
 
-        int secondPowerNumber = MathUtils.secondPower(10);
+        double secondPowerNumber = MathUtils.secondPower(10);
         System.out.println(secondPowerNumber);
     }
 }
